Delegate to the `Transformer` version of the `filter` method
+review REVIEW-6081
