+++ subprojects/core/src/main/java/org/gradle/api/internal/file/copy/ClosureBackedTransformer.java
@@ -1,0 +1,33 @@
+/*
+ * Copyright 2016 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.gradle.api.internal.file.copy;
+
+import groovy.lang.Closure;
+import org.gradle.api.Transformer;
+
+class ClosureBackedTransformer implements Transformer<String, String> {
+    private final Closure closure;
+
+    public ClosureBackedTransformer(Closure closure) {
+        this.closure = closure;
+    }
+
+    @Override
+    public String transform(String s) {
+        Object val = closure.call(s);
+        return val == null ? null : val.toString();
+    }
+}
